BL-2833 make cover thumbnail for epub
Previously we were depending on a thumbnail
that gets created during upload
